+++ src/appleseed/renderer/modeling/environmentshader/edfenvironmentshader.cpp
@@ -121,7 +121,7 @@ namespace
             // Evaluate and store alpha value.
             InputValues uniform_values;
             m_inputs.evaluate_uniforms(&uniform_values);
-            m_alpha_value = static_cast<float>(uniform_values.m_alpha_value);
+            m_alpha_value = uniform_values.m_alpha_value;
 
             return true;
         }
